requirements.txt
streamlit
joblib
easyocr
numpy
pillow
python-dotenv
google-generativeai
requests
scikit-learn